+++ RelationshipService.Domain/Entities/UserProfileHobby.cs
@@ -2,7 +2,7 @@
 
 public class UserProfileHobby
 {
-    public int UserProfileId { get; set; }
+    public long UserProfileId { get; set; }
     public UserProfile UserProfile { get; set; }
     public int HobbyId { get; set; }
     public Hobby Hobby { get; set; }
